backend/requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
pytesseract==0.3.10
SpeechRecognition==3.10.1
requests==2.31.0
python-dotenv==1.0.1 